Renamed folders to end in use case
