+++ api/ExpressedRealms.Powers.Repository/PowerPrerequisites/CreatePrerequisiteUseCase/ICreatePrerequisiteUseCase.cs
@@ -1,6 +1,6 @@
 using ExpressedRealms.Shared;
 using FluentResults;
 
-namespace ExpressedRealms.Powers.Repository.PowerPrerequisites.CreatePrerequisite;
+namespace ExpressedRealms.Powers.Repository.PowerPrerequisites.CreatePrerequisiteUseCase;
 
 public interface ICreatePrerequisiteUseCase : IGenericUseCase<Result, CreatePrerequisiteModel> { }
