--- api/ExpressedRealms.Powers.Repository/PowerPrerequisites/DeletePrerequisite/DeletePrerequisiteModel.cs
@@ -1,6 +1,0 @@
-namespace ExpressedRealms.Powers.Repository.PowerPrerequisites.DeletePrerequisite;
-
-public class DeletePrerequisiteModel
-{
-    public int Id { get; set; }
-}
