+++ api/ExpressedRealms.Powers.Repository/PowerPrerequisites/DeletePrerequisiteUseCase/DeletePrerequisiteModel.cs
@@ -1,0 +1,6 @@
+namespace ExpressedRealms.Powers.Repository.PowerPrerequisites.DeletePrerequisiteUseCase;
+
+public class DeletePrerequisiteModel
+{
+    public int Id { get; set; }
+}
