+++ api/ExpressedRealms.Powers.Repository/PowerPrerequisites/DeletePrerequisiteUseCase/DeletePrerequisiteUseCase.cs
@@ -1,6 +1,6 @@
 using FluentResults;
 
-namespace ExpressedRealms.Powers.Repository.PowerPrerequisites.DeletePrerequisite;
+namespace ExpressedRealms.Powers.Repository.PowerPrerequisites.DeletePrerequisiteUseCase;
 
 public class DeletePrerequisiteUseCase(
     DeletePrerequisiteModelValidator validator,
