+++ api/ExpressedRealms.Powers.Repository/PowerPrerequisites/DeletePrerequisiteUseCase/IDeletePrerequisiteUseCase.cs
@@ -1,6 +1,6 @@
 using ExpressedRealms.Shared;
 using FluentResults;
 
-namespace ExpressedRealms.Powers.Repository.PowerPrerequisites.DeletePrerequisite;
+namespace ExpressedRealms.Powers.Repository.PowerPrerequisites.DeletePrerequisiteUseCase;
 
 public interface IDeletePrerequisiteUseCase : IGenericUseCase<Result, DeletePrerequisiteModel> { }
