+++ api/ExpressedRealms.Powers.Repository/PowerPrerequisites/EditPrerequisiteUseCase/IEditPrerequisiteUseCase.cs
@@ -1,6 +1,6 @@
 using ExpressedRealms.Shared;
 using FluentResults;
 
-namespace ExpressedRealms.Powers.Repository.PowerPrerequisites.EditPrerequisite;
+namespace ExpressedRealms.Powers.Repository.PowerPrerequisites.EditPrerequisiteUseCase;
 
 public interface IEditPrerequisiteUseCase : IGenericUseCase<Result, EditPrerequisiteModel> { }
